+++ mkchromecast/__init__.py
@@ -2,7 +2,7 @@
 
 # This file is part of mkchromecast.
 
-from mkchromecast.audiodevices import *
+from mkchromecast.audio_devices import *
 import mkchromecast.colors as colors
 from mkchromecast.terminate import *
 from mkchromecast.version import __version__
@@ -167,6 +167,7 @@ Possible backends:
     - parec (default in Linux)
     - ffmpeg
     - avconv
+    - gstreamer
 
 Example:
     python mkchromecast.py --encoder-backend ffmpeg
@@ -230,6 +231,16 @@ Reboot the Google Cast device.
 )
 
 parser.add_argument(
+'--reconnect',
+action='store_true',
+default=False,
+help='''
+This flag monitors if connection with google cast has been lost, and try to
+reconnect.
+'''
+)
+
+parser.add_argument(
 '-s',
 '--select-cc',
 action='store_true',
@@ -274,6 +285,20 @@ Which sample rate to use?
     - 22050Hz: sampling rate of audio quality of AM radio.
 
 For more information see: http://wiki.audacityteam.org/wiki/Sample_Rates.
+'''
+)
+
+parser.add_argument(
+'--segment-time',
+type=int,
+default=None,
+help=
+'''
+Segmentate audio for improved live streaming when using ffmpeg.
+
+Example:
+    python mkchromecast.py --encoder-backend ffmpeg --segment-time 2
+
 '''
 )
 
@@ -407,6 +432,8 @@ if debug == True:
 discover = args.discover
 host = args.host
 sourceurl = args.source_url
+reconnect = args.reconnect
+
 ccname = args.name
 if debug == True:
     print('Google Cast name:', ccname)
@@ -479,9 +506,10 @@ if platform == 'Darwin':
 else:
     backends.remove('node')
     backends.append('parec')
+    backends.append('gstreamer')
 
 if args.debug == True:
-    print('backends: ',backends)
+    print('backends: ', backends)
 
 if args.encoder_backend not in backends and args.encoder_backend != None:
     print(colors.error('Supported backends are: '))
@@ -571,6 +599,19 @@ if args.sample_rate != 0:
         samplerate = abs(args.sample_rate)
 elif args.sample_rate == 0:
     samplerate = 44100
+
+"""
+Segment time
+"""
+avoid = ['parec', 'node']
+if isinstance(args.segment_time, int) and backend not in avoid:
+    segmenttime = args.segment_time
+elif isinstance(args.segment_time, float) or backend in avoid:
+    segmenttime = None
+else:
+    print(colors.warning('The segment time has to be an integer number'))
+    print(colors.warning('Set to default of 2 seconds'))
+    segmenttime = 2
 
 """
 Video