+++ mkchromecast/audio.py
@@ -7,7 +7,7 @@ Google Cast device has to point out to http://ip:5000/stream
 """
 
 import mkchromecast.__init__
-from mkchromecast.audiodevices import *
+from mkchromecast.audio_devices import *
 import mkchromecast.colors as colors
 from mkchromecast.config import *
 import mkchromecast.messages as msg
@@ -41,6 +41,8 @@ In this block we check variables from __init__.py
 tray = mkchromecast.__init__.tray
 adevice = mkchromecast.__init__.adevice
 chunk_size = mkchromecast.__init__.chunk_size
+segmenttime = mkchromecast.__init__.segmenttime
+
 if debug == True:
     print(':::audio::: chunk_size: ', chunk_size)
 debug = mkchromecast.__init__.debug
@@ -85,11 +87,14 @@ else:
         codec= ConfigSectionMap('settings')['codec']
         bitrate = ConfigSectionMap('settings')['bitrate']
         samplerate= ConfigSectionMap('settings')['samplerate']
+        adevice = ConfigSectionMap('settings')['alsadevice']
+        if adevice == 'None':
+            adevice = None
         if debug == True:
             print(':::audio::: tray ='+str(tray))
-            print(colors.warning('Configuration file exist'))
+            print(colors.warning('Configuration file exists'))
             print(colors.warning('Using defaults set there'))
-            print(backend,codec,bitrate,samplerate)
+            print(backend, codec, bitrate, samplerate, adevice)
     else:
         backend = mkchromecast.__init__.backend
         backends_dict[backend] = backend
@@ -146,12 +151,15 @@ else:
             if codec == 'mp3' and int(bitrate) > 320:
                 bitrate = '320'
                 msg.maxbitrate(codec, bitrate)
-
-            if codec == 'ogg' or codec == 'aac' and int(bitrate) > 500:
+            elif codec == 'ogg' and int(bitrate) > 500:
                 bitrate = '500'
                 msg.maxbitrate(codec, bitrate)
+            elif codec == 'aac' and int(bitrate) > 500:
+                bitrate = '500'
+                msg.maxbitrate(codec, bitrate)
+            else:
+                bitrate = bitrate+'k'
 
-            bitrate = bitrate+'k'
             if sourceurl == None:
                 print(colors.options('Selected bitrate:')+' '+ bitrate)
 
@@ -241,12 +249,18 @@ else:
         print (command)
         return
 
+    def set_segmenttime():
+        string = [ '-f', 'segment', '-segment_time', str(segmenttime) ]
+        for element in string:
+            command.insert(-9, element)
+        return
+
     """
     MP3 192k
     """
     if  codec == 'mp3':
 
-        if platform == 'Linux' and backends_dict[backend] != 'parec':
+        if platform == 'Linux' and backends_dict[backend] != 'parec' and backends_dict[backend] != 'gstreamer':
             command = [
                 backend,
                 '-ac', '2',
@@ -262,13 +276,44 @@ else:
                 ]
             if adevice != None:
                 modalsa()
-        elif platform == 'Linux' and backends_dict[backend] == 'parec':
+
+            if segmenttime != None:
+                set_segmenttime()
+
+        elif platform == 'Linux' and backends_dict[backend] == 'parec' or backends_dict[backend] == 'gstreamer':
             command = [
                 'lame',
                 '-b', bitrate[:-1],
                 '-r',
                 '-'
                 ]
+            """
+        This command dumps to file correctly, but does not work for stdout.
+        elif platform == 'Linux' and backends_dict[backend] == 'gstreamer':
+            command = [
+                'gst-launch-1.0',
+                '-v',
+                '!',
+                'audioconvert',
+                '!',
+                'audio/x-raw,rate='+samplerate,
+                '!',
+                'lamemp3enc',
+                'target=bitrate',
+                'bitrate='+bitrate[:-1],
+                'cbr=true',
+                '!',
+                'mpegaudioparse',
+                '!',
+                'filesink', 'location=/dev/stdout'
+                ]
+            if adevice != None:
+                command.insert(2, 'alsasrc')
+                command.insert(3, 'device="'+adevice+'"')
+            else:
+                command.insert(2, 'pulsesrc')
+                command.insert(3, 'device="mkchromecast.monitor"')
+            """
         else:
             command = [
                 backend,
@@ -282,14 +327,14 @@ else:
                 '-b:a', bitrate,
                 'pipe:'
                 ]
-        if debug == False and backends_dict[backend] != 'parec':
-            debug_command()
+            if segmenttime != None:
+                set_segmenttime()
 
     """
     OGG 192k
     """
     if  codec == 'ogg':
-        if platform == 'Linux' and backends_dict[backend] != 'parec':
+        if platform == 'Linux' and backends_dict[backend] != 'parec' and backends_dict[backend] != 'gstreamer':
             command = [
                 backend,
                 '-ac', '2',
@@ -305,7 +350,11 @@ else:
                 ]
             if adevice != None:
                 modalsa()
-        elif platform == 'Linux' and backends_dict[backend] == 'parec':
+
+            if segmenttime != None:
+                set_segmenttime()
+
+        elif platform == 'Linux' and backends_dict[backend] == 'parec' or backends_dict[backend] == 'gstreamer':
             command = [
                 'oggenc',
                 '-b', bitrate[:-1],
@@ -314,6 +363,33 @@ else:
                 '--ignorelength',
                 '-'
                 ]
+            """
+        This command dumps to file correctly, but does not work for stdout.
+        elif platform == 'Linux' and backends_dict[backend] == 'gstreamer':
+            command = [
+                'gst-launch-1.0',
+                '!',
+                'audioconvert',
+                '!',
+                'audioresample',
+                '!',
+                'vorbisenc',
+                #'bitrate='+str(int(bitrate[:-1])*1000),
+                '!',
+                'vorbisparse',
+                '!',
+                'oggmux',
+                '!',
+                'filesink', 'location=/dev/stdout'
+                #gst-launch-1.0 pulsesrc device="mkchromecast.monitor" ! audioconvert ! audioresample ! vorbisenc ! oggmux ! filesink
+                ]
+            if adevice != None:
+                command.insert(1, 'alsasrc')
+                command.insert(2, 'device="'+adevice+'"')
+            else:
+                command.insert(1, 'pulsesrc')
+                command.insert(2, 'device="mkchromecast.monitor"')
+            """
         else:
             command = [
                 backend,
@@ -321,20 +397,19 @@ else:
                 '-audio_device_index', '0',
                 '-i', '',
                 '-acodec', 'libvorbis',
+                '-f', 'segment', '-segment_time', '2',
                 '-f', 'ogg',
                 '-ac', '2',
                 '-ar', samplerate,
                 '-b:a', bitrate,
                 'pipe:'
                 ]
-        if debug == False and backends_dict[backend] != 'parec':
-            debug_command()
 
     """
     AAC > 128k for Stereo, Default sample rate: 44100kHz
     """
     if  codec == 'aac':
-        if platform == 'Linux' and backends_dict[backend] != 'parec':
+        if platform == 'Linux' and backends_dict[backend] != 'parec' and backends_dict[backend] != 'gstreamer':
             command = [
                 backend,
                 '-ac', '2',
@@ -351,7 +426,8 @@ else:
                 ]
             if adevice != None:
                 modalsa()
-        elif platform == 'Linux' and backends_dict[backend] == 'parec':
+
+        elif platform == 'Linux' and backends_dict[backend] == 'parec' or backends_dict[backend] == 'gstreamer':
             command = [
                 'faac',
                 '-b', bitrate[:-1],
@@ -362,6 +438,31 @@ else:
                 '-',
                 '-'
                 ]
+            """
+        This command dumps to file correctly, but does not work for stdout.
+        elif platform == 'Linux' and backends_dict[backend] == 'gstreamer':
+            command = [
+                'gst-launch-1.0',
+                '-v',
+                '!',
+                'audioconvert',
+                '!',
+                'audio/x-raw,rate='+samplerate,
+                '!',
+                'voaacenc',
+                #'bitrate='+bitrate[:-1],
+                '!',
+                'aacparse',
+                '!',
+                'filesink', 'location=/dev/stdout'
+                ]
+            if adevice != None:
+                command.insert(2, 'alsasrc')
+                command.insert(3, 'device="'+adevice+'"')
+            else:
+                command.insert(2, 'pulsesrc')
+                command.insert(3, 'device="mkchromecast.monitor"')
+            """
         else:
             command = [
                 backend,
@@ -376,8 +477,8 @@ else:
                 '-cutoff', '18000',
                 'pipe:'
                 ]
-        if debug == False and backends_dict[backend] != 'parec':
-            debug_command()
+            if segmenttime != None:
+                set_segmenttime()
 
     """
     WAV 24-Bit
@@ -398,7 +499,11 @@ else:
                 ]
             if adevice != None:
                 modalsa()
-        elif platform == 'Linux' and backends_dict[backend] == 'parec':
+
+            if segmenttime != None:
+                set_segmenttime()
+
+        elif platform == 'Linux' and backends_dict[backend] == 'parec' or backends_dict[backend] == 'gstreamer':
             command = [
                 'sox',
                 '-t', 'raw',
@@ -426,8 +531,8 @@ else:
                 '-ar', samplerate,
                 'pipe:'
                 ]
-        if debug == False and backends_dict[backend] != 'parec':
-            debug_command()
+            if segmenttime != None:
+                set_segmenttime()
 
     """
     FLAC 24-Bit (values taken from: https://trac.ffmpeg.org/wiki/Encode/HighQualityAudio) except for parec.
@@ -448,7 +553,11 @@ else:
                 ]
             if adevice != None:
                 modalsa()
-        elif platform == 'Linux' and backends_dict[backend] == 'parec':
+
+            if segmenttime != None:
+                set_segmenttime()
+
+        elif platform == 'Linux' and backends_dict[backend] == 'parec' or backends_dict[backend] == 'gstreamer':
             command = [
                 'flac',
                 '-',
@@ -472,8 +581,12 @@ else:
                 '-ar', samplerate,
                 'pipe:'
                 ]
-        if debug == False and backends_dict[backend] != 'parec':
-            debug_command()
+            if segmenttime != None:
+                set_segmenttime()
+
+    verbose_backend = ['ffmpeg', 'avconv']
+    if debug == False and backends_dict[backend] in verbose_backend:
+        debug_command()
 
 app = Flask(__name__)
 
@@ -509,8 +622,9 @@ def shutdown():
 """
 @app.route('/' + appendtourl)
 def stream():
-    if platform == 'Linux' and bool(backends_dict) == True \
-            and backends_dict[backend] == 'parec':
+    if (platform == 'Linux'
+            and bool(backends_dict) == True
+            and backends_dict[backend] == 'parec'):
         c_parec = [
             backend,
             '--format=s16le',
@@ -518,15 +632,39 @@ def stream():
             ]
         parec = Popen(c_parec, stdout=PIPE)
         process = Popen(command, stdin=parec.stdout, stdout=PIPE, bufsize=-1)
+    elif (platform == 'Linux' and
+            bool(backends_dict) == True and
+            backends_dict[backend] == 'gstreamer'):
+        c_gst = [
+            'gst-launch-1.0',
+            '-v',
+            '!',
+            'audioconvert',
+            '!',
+            'filesink', 'location=/dev/stdout'
+            ]
+        if adevice != None:
+            c_gst.insert(2, 'alsasrc')
+            c_gst.insert(3, 'device="'+adevice+'"')
+        else:
+            c_gst.insert(2, 'pulsesrc')
+            c_gst.insert(3, 'device="mkchromecast.monitor"')
+        gst = Popen(c_gst, stdout=PIPE)
+        process = Popen(command, stdin=gst.stdout, stdout=PIPE, bufsize=-1)
     else:
         process = Popen(command, stdout=PIPE, bufsize=-1)
     read_chunk = partial(os.read, process.stdout.fileno(), chunk_size)
     return Response(iter(read_chunk, b''), mimetype=mtype)
 
 def start_app():
+    """Starting the streaming server.
+
+    Note that passthrough_errors=False is useful when reconnecting. In that
+    way, flask won't die.
+    """
     monitor_daemon = monitor()
     monitor_daemon.start()
-    app.run(host= '0.0.0.0')
+    app.run(host= '0.0.0.0', passthrough_errors=False)
 
 class multi_proc(object):       # I launch ffmpeg in a different process
     def __init__(self):