+++ mkchromecast/cast.py
@@ -4,7 +4,7 @@
 
 from __future__ import print_function
 from mkchromecast.__init__ import *
-from mkchromecast.audiodevices import *
+from mkchromecast.audio_devices import *
 import mkchromecast.colors as colors
 from mkchromecast.config import *
 from mkchromecast.preferences import ConfigSectionMap
@@ -17,6 +17,8 @@ import os.path
 import pickle
 import subprocess
 import mkchromecast.colors as colors
+from threading import Thread
+
 """
 Configparser is imported differently in Python3
 """
@@ -26,6 +28,8 @@ except ImportError:
     import configparser as ConfigParser # This is for Python3
 
 class casting(object):
+    """Main casting class.
+    """
     def __init__(self):                     # __init__ to call the self.ip
         import mkchromecast.__init__        # This is to verify against some needed variables
         self.platform = mkchromecast.__init__.platform
@@ -38,6 +42,8 @@ class casting(object):
         self.discover = mkchromecast.__init__.discover
         self.host = mkchromecast.__init__.host
         self.ccname = mkchromecast.__init__.ccname
+        self.reconnect = mkchromecast.__init__.reconnect
+        self.title = 'Mkchromecast v'+mkchromecast.__init__.__version__
 
         if self.host == None:
             if self.platform == 'Linux':
@@ -66,7 +72,10 @@ class casting(object):
     def getnetworkip(self):
         s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
         s.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
-        s.connect(('<broadcast>', 0))
+        try:
+            s.connect(('<broadcast>', 0))
+        except socket.error:
+            self.ip = '127.0.0.1'
         self.discovered_ip = s.getsockname()[0]
         if self.debug == True:
             print(':::cast::: sockets method', self.discovered_ip)
@@ -84,6 +93,21 @@ class casting(object):
                     if self.debug == True:
                         print(':::cast::: netifaces method', self.discovered_ip)
 
+    def _get_chromecasts(self):
+        # compatibility
+        try:
+            return list(pychromecast.get_chromecasts_as_dict().keys())
+        except AttributeError:
+            self._chromecasts_by_name = {c.name: c for c in pychromecast.get_chromecasts()}
+            return list(self._chromecasts_by_name.keys())
+
+    def _get_chromecast(self, name):
+        # compatibility
+        try:
+            return pychromecast.get_chromecast(friendly_name=self.cast_to)
+        except AttributeError:
+            return self._chromecasts_by_name[name]
+
     """
     Cast processes
     """
@@ -91,7 +115,7 @@ class casting(object):
         import mkchromecast.__init__            # This is to verify against some needed variables.
         from pychromecast import socket_client  # This fixes the `No handlers could be found for logger "pychromecast.socket_client` warning"`.
                                                 # See commit 18005ebd4c96faccd69757bf3d126eb145687e0d.
-        self.cclist = list(pychromecast.get_chromecasts_as_dict().keys())
+        self.cclist = self._get_chromecasts()
         if self.debug == True:
             print('self.cclist', self.cclist)
 
@@ -123,8 +147,8 @@ class casting(object):
             if self.discover == False:
                 print(colors.important('We will cast to first device in the list above!'))
                 print(' ')
-                self.castto = self.cclist[0]
-                print(colors.success(self.castto))
+                self.cast_to = self.cclist[0]
+                print(colors.success(self.cast_to))
                 print(' ')
 
         elif len(self.cclist) != 0 and self.select_cc == True and self.tray == False and self.ccname == None:
@@ -143,17 +167,17 @@ class casting(object):
                 self.availablecc=[]
                 for self.index,device in enumerate(self.cclist):
                     print(str(self.index)+'      ',str(device))
-                    toappend = [self.index,device]
-                    self.availablecc.append(toappend)
+                    to_append = [self.index,device]
+                    self.availablecc.append(to_append)
                 """
             else:
                 if self.debug == True:
                     print('else:')
                 self.tf = open('/tmp/mkchromecast.tmp', 'rb')
                 self.index=pickle.load(self.tf)
-                self.castto = self.cclist[int(self.index)]
+                self.cast_to = self.cclist[int(self.index)]
                 print(' ')
-                print(colors.options('Casting to:')+' '+colors.success(self.castto))
+                print(colors.options('Casting to:')+' '+colors.success(self.cast_to))
                 print(' ')
 
         elif len(self.cclist) != 0 and self.select_cc == True and self.tray == True:
@@ -172,18 +196,18 @@ class casting(object):
                 self.availablecc=[]
                 for self.index,device in enumerate(self.cclist):
                     print(str(self.index)+'      ',str(device))
-                    toappend = [self.index,device]
-                    self.availablecc.append(toappend)
+                    to_append = [self.index,device]
+                    self.availablecc.append(to_append)
                 """
             else:
                 if self.debug == True:
                     print('else:')
                 self.tf = open('/tmp/mkchromecast.tmp', 'rb')
                 self.index=pickle.load(self.tf)
-                self.castto = self.cclist[int(self.index)]
+                self.cast_to = self.cclist[int(self.index)]
                 self.availablecc()
                 print(' ')
-                print(colors.options('Casting to:')+' '+colors.success(self.castto))
+                print(colors.options('Casting to:')+' '+colors.success(self.cast_to))
                 print(' ')
 
         elif len(self.cclist) == 0 and self.tray == False:
@@ -213,9 +237,9 @@ class casting(object):
             try:
                 pickle.dump(self.index, self.tf)
                 self.tf.close()
-                self.castto = self.cclist[int(self.index)]
+                self.cast_to = self.cclist[int(self.index)]
                 print(' ')
-                print(colors.options('Casting to:')+' '+colors.success(self.castto))
+                print(colors.options('Casting to:')+' '+colors.success(self.cast_to))
                 print(' ')
             except IndexError:
                 checkmktmp()
@@ -229,16 +253,16 @@ class casting(object):
             print('def get_cc(self):')
         try:
             if self.ccname != None:
-                self.castto = self.ccname
-            self.cast = pychromecast.get_chromecast(friendly_name=self.castto)
+                self.cast_to = self.ccname
+            self.cast = self._get_chromecast(self.cast_to)
             # Wait for cast device to be ready
             self.cast.wait()
             print(' ')
-            print(colors.important('Information about ')+' '+colors.success(self.castto))
+            print(colors.important('Information about ')+' '+colors.success(self.cast_to))
             print(' ')
             print(self.cast.device)
             print(' ')
-            print(colors.important('Status of device ')+' '+colors.success(self.castto))
+            print(colors.important('Status of device ')+' '+colors.success(self.cast_to))
             print(' ')
             print(self.cast.status)
             print(' ')
@@ -262,7 +286,7 @@ class casting(object):
             print('def play_cast(self):')
         localip = self.ip
 
-        print(colors.options('The IP of ')+colors.success(self.castto)+colors.options(' is:')+' '+self.cast.host)
+        print(colors.options('The IP of ')+colors.success(self.cast_to)+colors.options(' is:')+' '+self.cast.host)
         if self.host == None:
             print(colors.options('Your local IP is:')+' '+localip)
         else:
@@ -288,7 +312,7 @@ class casting(object):
             yt.play_video(video)
         else:
         """
-        ncast = self.cast
+        media_controller = self.cast.media_controller
 
         if self.tray == True:
             config = ConfigParser.RawConfigParser()
@@ -297,12 +321,12 @@ class casting(object):
 
             if os.path.exists(configf) and self.tray == True:
                 print(tray)
-                print(colors.warning('Configuration file exist'))
+                print(colors.warning('Configuration file exists'))
                 print(colors.warning('Using defaults set there'))
                 config.read(configf)
                 self.backend = ConfigSectionMap('settings')['backend']
 
-        if self.backend == 'ffmpeg' or self.backend == 'avconv' or self.backend == 'parec' and self.sourceurl == None:
+        if self.sourceurl != None:
             if args.video == True:
                 import mkchromecast.video
                 mtype = mkchromecast.video.mtype
@@ -310,28 +334,31 @@ class casting(object):
                 import mkchromecast.audio
                 mtype = mkchromecast.audio.mtype
             print(' ')
-            print(colors.options('The media type string used is:')+' '+mtype)
-            ncast.play_media('http://'+localip+':5000/stream', mtype)
-        elif self.sourceurl != None:
+            print(colors.options('Casting from stream URL:')+' '+self.sourceurl)
+            print(colors.options('Using media type:')+' '+mtype)
+            media_controller.play_media(self.sourceurl, mtype, title = self.title)
+        elif self.backend == 'ffmpeg' or self.backend == 'avconv' or self.backend == 'parec' or self.backend == 'gstreamer' and self.sourceurl == None:
             import mkchromecast.audio
             mtype = mkchromecast.audio.mtype
             print(' ')
-            print(colors.options('Casting from stream URL:')+' '+self.sourceurl)
-            print(colors.options('Using media type:')+' '+mtype)
-            ncast.play_media(self.sourceurl, mtype)
+            print(colors.options('The media type string used is:')+' '+mtype)
+            media_controller.play_media('http://'+localip+':5000/stream', mtype, title = self.title)
         else:
             print(' ')
             print(colors.options('The media type string used is:')+' '+  'audio/mpeg')
-            ncast.play_media('http://'+localip+':3000/stream.mp3', 'audio/mpeg')
+            media_controller.play_media('http://'+localip+':3000/stream.mp3', 'audio/mpeg', title = self.title)
         print(' ')
         print(colors.important('Cast media controller status'))
         print(' ')
-        print(ncast.status)
+        print(self.cast.status)
         print(' ')
+        if self.reconnect == True:
+            self.r = Thread(target = self.reconnect_cc)
+            self.r.daemon = True   # This has to be set to True so that we catch KeyboardInterrupt.
+            self.r.start()
 
     def stop_cast(self):
-        ncast = self.cast
-        ncast.quit_app()
+        self.cast.quit_app()
 
     def volume_up(self):
         """ Increment volume by 0.1 unless it is already maxed.
@@ -339,9 +366,8 @@ class casting(object):
         """
         print('Increasing volume...')
         print(' ')
-        ncast = self.cast
-        volume = round(ncast.status.volume_level, 1)
-        return ncast.set_volume(volume + 0.1)
+        volume = round(self.cast.status.volume_level, 1)
+        return self.cast.set_volume(volume + 0.1)
 
     def volume_down(self):
         """ Decrement the volume by 0.1 unless it is already 0.
@@ -349,13 +375,12 @@ class casting(object):
         """
         print('Decreasing volume...')
         print(' ')
-        ncast = self.cast
-        volume = round(ncast.status.volume_level, 1)
-        return ncast.set_volume(volume - 0.1)
+        volume = round(self.cast.status.volume_level, 1)
+        return self.cast.set_volume(volume - 0.1)
 
     def reboot(self):
         if self.platform == 'Darwin':
-            self.cast.host = socket.gethostbyname(self.castto+'.local')
+            self.cast.host = socket.gethostbyname(self.cast_to+'.local')
             reboot(self.cast.host)
         else:
             print(colors.error('This method is not supported in Linux yet.'))
@@ -370,5 +395,58 @@ class casting(object):
                 print(str(self.index)+'      ', str(device))
             except UnicodeEncodeError:
                 print(str(self.index)+'      ', str(unicode(device).encode("utf-8")))
-            toappend = [self.index,device]
-            self.availablecc.append(toappend)
+            to_append = [self.index,device]
+            self.availablecc.append(to_append)
+
+    def reconnect_cc(self):
+        """Dummy method to call  _reconnect_cc_().
+
+        In the cast that the self.r thread is alive, we check that the
+        chromecast is connected. If it is connected, we check again in
+        5 seconds.
+        """
+        try:
+            while self.r.is_alive():
+                self._reconnect_cc_()
+                time.sleep(5)       #FIXME: I think that this has to be set by users.
+        except KeyboardInterrupt:
+            self.stop_cast()
+            if platform == 'Darwin':
+                inputint()
+                outputint()
+            elif platform == 'Linux' and adevice == None:
+                from mkchromecast.pulseaudio import remove_sink
+                remove_sink()
+            terminate()
+
+    def _reconnect_cc_(self):
+        """Check if chromecast is disconnected and reconnect.
+
+        This function checks if the chromecast is online. Then, if the display
+        name is different from "Default Media Receiver", it reconnects to the
+        chromecast.
+        """
+        ip = self.cast.host
+        if ping_chromecast(ip) == True:     # The chromecast is online.
+            if str(self.cast.status.display_name) != "Default Media Receiver":
+                self.ccname = self.cast_to
+                self.get_cc()
+                self.play_cast()
+        else:                               # The chromecast is offline.
+            try:
+                self.ccname = self.cast_to
+                self.get_cc()
+                self.play_cast()
+            except AttributeError:
+                pass
+
+def ping_chromecast(ip):
+    """This function pings to hosts.
+
+    Credits: http://stackoverflow.com/a/34455969/1995261
+    """
+    try:
+        output = subprocess.check_output("ping -c 1 "+ip, shell=True)
+    except:
+        return False
+    return True